move SurfaceFigure tests from test_surf_plotting to displays.tests.test_surface_figure
